Mindestanforderungen Test:
Folgende Backend-Test wurden implementiert:
- testPostInventoryObject (Rot - Wirft noch einen Fehler)
- testPutInventoryObject (Rot - Methode in der Service-Klasse muss noch implementiert werden)
- testDeleteInventoryObject (Grün)

Die Delete-Methode ind er Service-Klasse wurde um einen Return-Wert ergänzt

+++ src/main/java/htwberlin/webtech/InventoryObjectService.java
@@ -29,10 +29,14 @@ public class InventoryObjectService {
     }
 
     public void update(Long id, int amount) {
-        //repo.findById(id).setAmount(amount);
+        // Muss noch implementiert werden;
     }
 
-    public void delete(Long id){
-        repo.deleteById(id);
+    public boolean delete(Long id) {
+        if (repo.existsById(id)) {
+            repo.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
